Fix an intermittent segfault
It was caused by comments being deleted, making the reported comment
count on the front page inaccurate (an overestimate). Therefore, the
entire comment array wasn't being filled, leaving a bunch of null
pointers to ruin everything.

There is still one more segfault that happens when newlines exist in a
comment (not <p>, but actual newlines in the HTML). This complicates
parsing and is sufficiently annoying that I'll leave an example here:
https://news.ycombinator.com/item?id=14460777