Add access to other sorting methods (new, etc)
I honestly did not expect this to work without modifications to the
parsing function, but hey.

Warning: I got a random segfault when pressing 'q' after switching tabs
for a while, but I can't reproduce it, so... I'll fix that in a later
commit when I find out what it actually is.